又修复了各种BUG

+++ user/personalfiles/inc/folder.php
@@ -17,20 +17,23 @@ http://dcms-social.ru
 $set['title'] = text($dir['name']);
 title();
 aut();
-// Редактирование и удаление файлов\папок    
+
+// 编辑和删除文件/文件夹
 if (isset($user) && (user_access('down_file_edit') || $ank['id'] == $user['id'])) {
-	// Удаление папок и файлов в них
+	// 删除其中的文件夹和文件
 	include "inc/folder.delete.php";
-	// Управление папками
+	// 文件夹管理
 	include "inc/folder.edit.php";
-	// Прочие формы вывода
+	// 其他形式的提款
 	include "inc/all.form.php";
 }
-// Вывод обратной навигации
+
+// 反向导航输出
 echo "<div class='foot'>";
 echo "<img src='/style/icons/up_dir.gif' alt='*'> " . ($dir['osn'] == 1 ? '文件' : '') . " " . user_files($dir['id_dires']) . " " . ($dir['osn'] == 1 ? '' : '&gt; ' . text($dir['name'])) . "";
 echo "</div>";
-// Перемещение файла в другую папку
+
+// 将文件移动到另一个文件夹
 if (isset($_GET['go']) && dbresult(dbquery("SELECT COUNT(*) FROM `downnik_files` WHERE `id` = '" . intval($_GET['go']) . "'"), 0) == 1) {
 	$file_go = dbassoc(dbquery("SELECT * FROM `downnik_files` WHERE `id` = '" . intval($_GET['go']) . "'"));
 	if (isset($_GET['ok']) && isset($_GET['ok']) && $ank['id'] == $user['id']) {
@@ -40,7 +43,8 @@ if (isset($_GET['go']) && dbresult(dbquery("SELECT COUNT(*) FROM `downnik_files`
 		exit;
 	}
 }
-/*--------------------Папка под паролем--------------------*/
+
+/*--------------------密码文件夹--------------------*/
 if ($dir['pass'] != NULL) {
 	if (isset($_POST['password'])) {
 		$_SESSION['pass'] = my_esc($_POST['password']);
@@ -52,14 +56,14 @@ if ($dir['pass'] != NULL) {
 	}
 	if (!user_access('down_dir_edit') && ($user['id'] != $ank['id'] && $_SESSION['pass'] != $dir['pass'])) {
 		echo '<form action="?" method="POST">密码: <br />		<input type="pass" name="password" value="" /><br />		
-<input type="submit" value="登录"/></form>';
+		      <input type="submit" value="登录"/></form>';
 		echo "<div class='foot'>";
 		echo "<img src='/style/icons/up_dir.gif' alt='*'> " . ($dir['osn'] == 1 ? '档案' : '') . " " . user_files($dir['id_dires']) . " " . ($dir['osn'] == 1 ? '' : '&gt; ' . text($dir['name'])) . "";
 		echo "</div>";
 		include_once '../../sys/inc/tfoot.php';
-		exit;
 	}
 }
+
 /*---------------------------------------------------------*/
 if (isset($_GET['go'])) {
 	echo '<div class="foot">';
@@ -69,6 +73,7 @@ if (isset($_GET['go'])) {
 	echo "选择文件的文件夹";
 	echo "</div>";
 }
+
 if (isset($_SESSION['down_dir']) || isset($_GET['down_dir'])) {
 	if (!isset($_SESSION['down_dir']) && dbresult(dbquery("SELECT COUNT(*) FROM `downnik_dir` WHERE `id` = '" . intval($_GET['down_dir']) . "' AND `upload` = '1'"), 0) == 1)
 		$_SESSION['down_dir'] = abs(intval($_GET['down_dir']));
@@ -78,6 +83,7 @@ if (isset($_SESSION['down_dir']) || isset($_GET['down_dir'])) {
 		echo "</div>";
 	}
 }
+
 $k_files = dbresult(dbquery("SELECT COUNT(*) FROM `downnik_files`  WHERE `my_dir` = '$dir[id]' AND `id_user` = '$ank[id]'"), 0);
 $k_post = dbresult(dbquery("SELECT COUNT(*) FROM `user_files` WHERE `id_dir` = '$dir[id]' AND `id_user` = '$ank[id]'"), 0);
 $k_post = $k_post + $k_files;
@@ -102,11 +108,11 @@ while ($post = dbassoc($q)) {
 	}
 	/*---------------------------*/
 	echo "<img src='/style/themes/$set[set_them]/loads/14/" . ($post['pass'] != null ? 'lock.gif' : 'dir.png') . "' alt='*'>";
-	if (isset($_GET['go'])) // Если перемещаем файл
+	if (isset($_GET['go'])) // 如果我们移动一个文件
 		echo " <a href='/user/personalfiles/$ank[id]/$post[id]/?go=$file_go[id]'>" . text($post['name']) . "</a>";
 	else
 		echo " <a href='/user/personalfiles/$ank[id]/$post[id]/'>" . text($post['name']) . "</a>";
-	/*----------------------Счетчик папок---------------------*/
+	/*----------------------文件夹计数器---------------------*/
 	$k_f = 0;
 	$q3 = dbquery("SELECT * FROM `user_files` WHERE `id_dires` like '%$post[id]%'");
 	while ($post2 = dbassoc($q3)) {
@@ -114,7 +120,7 @@ while ($post = dbassoc($q)) {
 	}
 	$k_f = $k_f + dbresult(dbquery("SELECT COUNT(*) FROM `user_files` WHERE `id_dir` = '$post[id]'"), 0);
 	/*---------------------------------------------------------*/
-	/*----------------------Счетчик файлов--------------------*/
+	/*----------------------文件计数器--------------------*/
 	$k_f2 = 0;
 	$q4 = dbquery("SELECT * FROM `user_files` WHERE `id_dires` like '%$post[id]%'");
 	while ($post3 = dbassoc($q4)) {
@@ -127,6 +133,7 @@ while ($post = dbassoc($q)) {
 		echo "<a href='?edit_folder=$post[id]'><img src='/style/icons/edit.gif' alt='*'></a> <a href='?delete_folder=$post[id]'><img src='/style/icons/delete.gif' alt='*'></a><br />";
 	echo "</div>";
 }
+
 if (!isset($_GET['go'])) {
 	$q2 = dbquery("SELECT * FROM `downnik_files`  WHERE `my_dir` = '$dir[id]' AND `id_user` = '$ank[id]' ORDER BY time DESC LIMIT $start, $set[p_str]");
 	//echo "<form method='post' action='?move_file'>";
@@ -158,9 +165,11 @@ if (!isset($_GET['go'])) {
 		else $ras = NULL;
 		echo '<a href="?id_file=' . $post['id'] . '&amp;page=' . $page . '"><b>' . text($post['name']) . '.' . $ras . '</b></a> (' . size_file($post['size']) . ') ';
 		if ($post['metka'] == 1) echo ' <font color=red>(18+)</font>';
-		if ($user['id'] == $post['id_user'] && $dir_id['my'] == 1) echo '<a href="/down/?trans=' . $post['id'] . '"><img src="/style/icons/z.gif" alt="*"> 到区域</a> ';
-		if (user_access('down_file_edit') || $user['id'] == $post['id_user']) echo '<a href="?id_file=' . $post['id'] . '&amp;edit"><img src="/style/icons/edit.gif" alt="*"></a> ';
-		if (user_access('down_file_delete') || $user['id'] == $post['id_user']) echo '<a href="?id_file=' . $post['id'] . '&amp;delete&amp;page=' . $page . '"><img src="/style/icons/delete.gif" alt="*"></a> ';
+		if (isset($user)) {
+			if ($user['id'] == $post['id_user'] && $dir_id['my'] == 1) echo '<a href="/down/?trans=' . $post['id'] . '"><img src="/style/icons/z.gif" alt="*"> 到区域</a> ';
+			if (user_access('down_file_edit') || $user['id'] == $post['id_user']) echo '<a href="?id_file=' . $post['id'] . '&amp;edit"><img src="/style/icons/edit.gif" alt="*"></a> ';
+			if (user_access('down_file_delete') || $user['id'] == $post['id_user']) echo '<a href="?id_file=' . $post['id'] . '&amp;delete&amp;page=' . $page . '"><img src="/style/icons/delete.gif" alt="*"></a> ';
+		}
 		echo '<br />';
 		if ($post['opis']) {
 			echo rez_text(text($post['opis'])) . '<br />';
@@ -169,10 +178,12 @@ if (!isset($_GET['go'])) {
 		echo '</div>';
 	}
 }
+
 //echo "<input value=\"任务\" type=\"submit\" name=\"job\" />";
 //echo "</form>";
 echo "</table>";
 if ($k_page > 1) str('?', $k_page, $page); // 输出页数
+
 echo "<div class='foot'>";
 echo "<img src='/style/icons/up_dir.gif' alt='*'> " . ($dir['osn'] == 1 ? '文件' : '') . " " . user_files($dir['id_dires']) . " " . ($dir['osn'] == 1 ? '' : '&gt; ' . text($dir['name'])) . "";
 echo "</div>";
